requested fixes
