+++ Geist Heist/Assets/Scenes/Milestone 1/Level 1.unity
@@ -759,6 +759,10 @@ PrefabInstance:
       value: Globe
       objectReference: {fileID: 0}
     - target: {fileID: 7413776157936059847, guid: c75e04841a4376b49bead083818ef1d6, type: 3}
+      propertyPath: interactableCanvas
+      value: 
+      objectReference: {fileID: 187208572}
+    - target: {fileID: 7413776157936059847, guid: c75e04841a4376b49bead083818ef1d6, type: 3}
       propertyPath: raycastLayer.m_Bits
       value: 128
       objectReference: {fileID: 0}
