+++ Geist Heist/Assets/Scenes/Milestone 1/Lobby.unity
@@ -898,6 +898,10 @@ PrefabInstance:
       value: Level 1
       objectReference: {fileID: 0}
     - target: {fileID: 7413776157936059847, guid: c75e04841a4376b49bead083818ef1d6, type: 3}
+      propertyPath: interactableCanvas
+      value: 
+      objectReference: {fileID: 193024564}
+    - target: {fileID: 7413776157936059847, guid: c75e04841a4376b49bead083818ef1d6, type: 3}
       propertyPath: raycastLayer.m_Bits
       value: 128
       objectReference: {fileID: 0}
@@ -1410,11 +1414,6 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: c81217ad42a978241917009c21da5118, type: 3}
   m_Name: 
   m_EditorClassIdentifier: Assembly-CSharp::SceneTransition
-  rayLength: 15
-  raycastLayer:
-    serializedVersion: 2
-    m_Bits: 128
-  canvas: {fileID: 193024564}
 --- !u!1 &2048236346
 GameObject:
   m_ObjectHideFlags: 0
