--- Geist Heist/Assets/Scripts/Player/Movement/FirstPersonInputHandler.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 87706826846a65d42a9390a9728aaa8c
